+++ js/components/dataset/cards-form.vue
@@ -1,4 +1,6 @@
 <style lang="less">
+@card-padding: 8px;
+
 .completer-row {
     @field-height: 34px;
     margin-bottom: 10px;
@@ -21,16 +23,18 @@
 
 
     .dataset-card-container {
+        margin-bottom: 10px;
+
         button.close {
-            position: relative;
-            right: 10px;
+            position: absolute;
+            right: @card-padding + 12px;
             top: 5px;
             z-index: 15;
             color: red;
             opacity: 1;
 
             &:hover {
-                right: 8px;
+                right: @card-padding + 10px;
                 top: 2px;
                 font-size: 2em;
             }
@@ -66,8 +70,8 @@
     <div class="row" v-show="!datasets.length">
         <p class="lead text-center">{{ _('No related datasets') }}</p>
     </div>
-    <div class="row" v-el:sortable v-show="datasets.length">
-        <div class="col-md-6 dataset-card-container"
+    <div class="card-list card-list--columned" v-el:sortable v-show="datasets.length">
+        <div class="col-xs-12 col-md-6 col-lg-4 dataset-card-container"
             v-for="datasetid in datasets | ids"
             :data-id="datasetid">
             <button type="button" class="close" @click="on_remove(datasetid)">