+++ js/components/user/card.vue
@@ -1,20 +1,15 @@
 <template>
-<div class="card user-card"
-    :class="{ 'pointer': clickable, 'selected': selected }" @click="click">
-    <a class="card-logo">
+<a class="card user-card" :class="{ selected: selected }" :title="user | display"
+    :href="clickable" @click.prevent="click">
+    <div class="card-logo">
         <img :alt="user | display" :src="user | avatar_url 60">
-    </a>
+    </div>
     <div class="card-body">
-        <h4>
-            <a :title="user | display">
-                {{ user | display }}
-            </a>
-        </h4>
-
+        <h4>{{ user | display }}</h4>
         <div class="clamp-3">{{{ user.about | markdown 180 }}}</div>
     </div>
 
-    <footer v-if="user.metrics" class="card__footer">
+    <footer v-if="user.metrics" class="card-footer">
         <ul>
             <li v-tooltip :title="_('Datasets')">
                 <span class="fa fa-cubes fa-fw"></span>
@@ -30,7 +25,7 @@
             </li>
         </ul>
     </footer>
-</div>
+</a>
 </template>
 
 <script>
@@ -45,7 +40,7 @@ export default {
         userid: null,
         clickable: {
             type: Boolean,
-            default: true
+            default: false
         },
         selected: {
             type: Boolean,